Use given WindowTreeHost instead of creating a new DesktopWindowTreeHost
On embed, WindowTreeClient creates a WindowTreeHost and passes it to
MusClient. Previously, MusClient ignored it, and so when MusClient
created a new widget, it would create a new DesktopWindowTreeHost,
creating a new top level window in the process. Now on embed, MusClient
uses the WindowTreeHost it is given instead of creating a new
DesktopWindowTreeHost, and our widget essentially ignores everything
related to DesktopWindowTreeHost. No new top level window is created so
we don't need the invalid user id hacks anymore and nothing shows up
when quick_launch launches.